[REFACTOR] User 관련 패키지 정리
