+++ src/main/java/com/triprecord/triprecord/record/service/RecordService.java
@@ -1,15 +1,10 @@
 package com.triprecord.triprecord.record.service;
 
-import com.triprecord.triprecord.global.exception.ErrorCode;
-import com.triprecord.triprecord.global.exception.TripRecordException;
 import com.triprecord.triprecord.record.entity.Record;
-import com.triprecord.triprecord.record.entity.RecordPlace;
-import com.triprecord.triprecord.record.repository.RecordPlaceRepository;
 import com.triprecord.triprecord.record.repository.RecordRepository;
 import com.triprecord.triprecord.record.controller.request.RecordCreateRequest;
-import com.triprecord.triprecord.user.UserService;
+import com.triprecord.triprecord.user.service.UserService;
 import com.triprecord.triprecord.user.entity.User;
-import com.triprecord.triprecord.user.repository.UserRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
