+++ src/main/java/com/triprecord/triprecord/user/controller/UserController.java
@@ -1,6 +1,9 @@
-package com.triprecord.triprecord.user;
+package com.triprecord.triprecord.user.controller;
 
-import jakarta.security.auth.message.AuthException;
+import com.triprecord.triprecord.user.controller.dto.request.UserCreateRequest;
+import com.triprecord.triprecord.user.controller.dto.response.UserInfoResponse;
+import com.triprecord.triprecord.user.controller.dto.request.UserLoginRequest;
+import com.triprecord.triprecord.user.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
