+++ src/main/java/com/triprecord/triprecord/user/controller/dto/request/UserCreateRequest.java
@@ -1,4 +1,4 @@
-package com.triprecord.triprecord.user;
+package com.triprecord.triprecord.user.controller.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
